+++ source/dataops/airflow/dags/dataops_e2e_dag.py
@@ -11,16 +11,19 @@ Notes:
 from __future__ import annotations
 
 import os
-import subprocess
 import sys
 from datetime import datetime
+from functools import partial
 from pathlib import Path
 
 from airflow.decorators import dag, task
 
-
 # source/dataops/airflow/dags/us8_dataops_e2e_dag.py -> repo root at parents[4]
 PROJECT_ROOT = Path(os.getenv("PROJECT_ROOT", Path(__file__).resolve().parents[4]))
+sys.path.insert(0, str(PROJECT_ROOT))
+
+from source.common.dag_utils import run_python_script as _run_script
+
 INGEST_SCRIPT = PROJECT_ROOT / "source" / "dataops" / "load_raw_data.py"
 CLEANSE_SCRIPT = PROJECT_ROOT / "source" / "dataops" / "cleanse_data.py"
 TRANSFORM_SCRIPT = PROJECT_ROOT / "source" / "dataops" / "build_customer_features.py"
@@ -28,17 +31,7 @@ LINEAGE_SCRIPT = PROJECT_ROOT / "source" / "dataops" / "generate_lineage.py"
 EDA_SCRIPT = PROJECT_ROOT / "source" / "dataops" / "run_eda.py"
 EDA_IMAGES_REPORT_SCRIPT = PROJECT_ROOT / "source" / "dataops" / "generate_eda_images_report.py"
 
-
-def run_python_script(script_path: Path) -> None:
-    """Run one project script and fail the task on non-zero exit code."""
-    if not script_path.exists():
-        raise FileNotFoundError(f"Script not found: {script_path}")
-
-    subprocess.run(
-        [sys.executable, "-B", str(script_path)],
-        cwd=str(PROJECT_ROOT),
-        check=True,
-    )
+run_python_script = partial(_run_script, cwd=PROJECT_ROOT)
 
 
 @dag(
