+++ source/dataops/airflow/dags/transform_and_lineage_dag.py
@@ -16,32 +16,26 @@ Environment variables used by downstream scripts:
 """
 
 from __future__ import annotations
+
 import os
-import subprocess
 import sys
 from datetime import datetime
+from functools import partial
 from pathlib import Path
 
 from airflow.decorators import dag, task
 
-
 # If PROJECT_ROOT is not provided, resolve from this file location:
 # source/dataops/airflow/dags/transform_and_lineage_dag.py -> repo root at parents[4]
 PROJECT_ROOT = Path(os.getenv("PROJECT_ROOT", Path(__file__).resolve().parents[4]))
+sys.path.insert(0, str(PROJECT_ROOT))
+
+from source.common.dag_utils import run_python_script as _run_script
+
 TRANSFORM_SCRIPT = PROJECT_ROOT / "source" / "dataops" / "build_customer_features.py"
 LINEAGE_SCRIPT = PROJECT_ROOT / "source" / "dataops" / "generate_lineage.py"
 
-
-def run_python_script(script_path: Path) -> None:
-    """Run one project script with the current interpreter and fail task on non-zero exit."""
-    if not script_path.exists():
-        raise FileNotFoundError(f"Script not found: {script_path}")
-
-    subprocess.run(
-        [sys.executable, "-B", str(script_path)],
-        cwd=str(PROJECT_ROOT),
-        check=True,
-    )
+run_python_script = partial(_run_script, cwd=PROJECT_ROOT)
 
 
 @dag(
