+++ source/dataops/airflow/dags/weekly_model_monitoring_dag.py
@@ -18,6 +18,8 @@ from __future__ import annotations
 import hashlib
 import json
 import os
+import re
+import sys
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
 from typing import Any
@@ -39,73 +41,14 @@ DEFAULT_TOP_5_FEATURES = [
 ]
 
 PROJECT_ROOT = Path(os.getenv("PROJECT_ROOT", Path(__file__).resolve().parents[4]))
+sys.path.insert(0, str(PROJECT_ROOT))
 PERM_IMPORTANCE_PATH = PROJECT_ROOT / "docs" / "artifacts" / "permutation_importance.csv"
 BEST_MODEL_PATH = PROJECT_ROOT / "docs" / "artifacts" / "us10_best_model.json"
 REGISTRY_EVIDENCE_PATH = PROJECT_ROOT / "docs" / "artifacts" / "us12_model_registry.json"
 
 
-def get_pg_conn():
-    return psycopg2.connect(
-        host=os.getenv("POSTGRES_HOST", "127.0.0.1"),
-        port=int(os.getenv("POSTGRES_PORT", "5432")),
-        dbname=os.getenv("POSTGRES_DB", "kkbox"),
-        user=os.getenv("POSTGRES_USER", "bt4301"),
-        password=os.getenv("POSTGRES_PASSWORD", "bt4301pass"),
-    )
-
-
-def roc_auc_binary(y_true: np.ndarray, y_score: np.ndarray) -> float:
-    y_true = np.asarray(y_true, dtype=np.int64)
-    y_score = np.asarray(y_score, dtype=np.float64)
-    pos = np.sum(y_true == 1)
-    neg = np.sum(y_true == 0)
-
-    if pos == 0 or neg == 0:
-        raise ValueError("AUC needs both positive and negative classes.")
-
-    order = np.argsort(-y_score, kind="mergesort")
-    y_sorted = y_true[order]
-    s_sorted = y_score[order]
-
-    tps = np.cumsum(y_sorted == 1)
-    fps = np.cumsum(y_sorted == 0)
-    distinct_idx = np.where(np.diff(s_sorted))[0]
-    threshold_idx = np.r_[distinct_idx, y_sorted.size - 1]
-
-    tpr = np.r_[0.0, tps[threshold_idx] / pos, 1.0]
-    fpr = np.r_[0.0, fps[threshold_idx] / neg, 1.0]
-    return float(np.trapz(tpr, fpr))
-
-
-def compute_psi(baseline: np.ndarray, current: np.ndarray, bins: int = 10) -> float:
-    baseline = np.asarray(baseline, dtype=np.float64)
-    current = np.asarray(current, dtype=np.float64)
-    baseline = baseline[np.isfinite(baseline)]
-    current = current[np.isfinite(current)]
-
-    if baseline.size < 2 or current.size < 2:
-        return float("nan")
-
-    quantiles = np.linspace(0.0, 1.0, bins + 1)
-    edges = np.quantile(baseline, quantiles)
-    edges = np.unique(edges)
-
-    if edges.size < 2:
-        return 0.0
-
-    # Ensure all out-of-range current values are counted in tail bins.
-    edges = edges.astype(np.float64)
-    edges[0] = -np.inf
-    edges[-1] = np.inf
-
-    baseline_hist, _ = np.histogram(baseline, bins=edges)
-    current_hist, _ = np.histogram(current, bins=edges)
-
-    eps = 1e-6
-    baseline_pct = (baseline_hist / max(1, baseline_hist.sum())).astype(np.float64) + eps
-    current_pct = (current_hist / max(1, current_hist.sum())).astype(np.float64) + eps
-
-    return float(np.sum((current_pct - baseline_pct) * np.log(current_pct / baseline_pct)))
+from source.common.db import get_connection as get_pg_conn
+from source.common.monitoring_utils import compute_psi, roc_auc_binary
 
 
 def stable_bucket(key: str) -> int:
@@ -192,8 +135,13 @@ def us14_weekly_model_monitoring():
         baseline_start = now_utc - timedelta(days=90)
         min_rows = 50
 
+        _SAFE_COLUMN_RE = re.compile(r"^[a-zA-Z0-9_]+$")
+
         with get_pg_conn() as conn:
             top_5_features = resolve_top_monitor_features(conn)
+            for feat in top_5_features:
+                if not _SAFE_COLUMN_RE.match(feat):
+                    raise ValueError(f"Unsafe feature name rejected: {feat!r}")
             with conn.cursor(cursor_factory=RealDictCursor) as cur:
                 cur.execute(
                     """
