+++ source/dataops/build_customer_features.py
@@ -14,17 +14,18 @@ Environment variables (matches docker-compose.yml / .env):
 """
 
 import os
+import sys
 import time
+from pathlib import Path
+
+_PROJECT_ROOT = Path(__file__).resolve().parents[2]
+sys.path.insert(0, str(_PROJECT_ROOT))
 
 import psycopg2
 
-DB_CONFIG = {
-    "host":     os.getenv("POSTGRES_HOST",     "localhost"),
-    "port":     int(os.getenv("POSTGRES_PORT", 5432)),
-    "dbname":   os.getenv("POSTGRES_DB",       "kkbox"),
-    "user":     os.getenv("POSTGRES_USER",     "bt4301"),
-    "password": os.getenv("POSTGRES_PASSWORD", "bt4301pass"),
-}
+from source.common.db import get_db_config
+
+DB_CONFIG = get_db_config()
 
 DDL = """
 CREATE TABLE IF NOT EXISTS processed.customer_features (
