+++ source/dataops/load_raw_data.py
@@ -11,16 +11,17 @@ Environment variables (matches docker-compose.yml / .env):
 """
 
 import os
+import sys
+from pathlib import Path
+
+_PROJECT_ROOT = Path(__file__).resolve().parents[2]
+sys.path.insert(0, str(_PROJECT_ROOT))
 
 import psycopg2
 
-DB_CONFIG = {
-    "host":     os.getenv("POSTGRES_HOST",     "localhost"),
-    "port":     int(os.getenv("POSTGRES_PORT", 5432)),
-    "dbname":   os.getenv("POSTGRES_DB",       "kkbox"),
-    "user":     os.getenv("POSTGRES_USER",     "bt4301"),
-    "password": os.getenv("POSTGRES_PASSWORD", "bt4301pass"),
-}
+from source.common.db import get_db_config
+
+DB_CONFIG = get_db_config()
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), "..", "..", "data", "raw")
