+++ source/mlops/feature_selection.py
@@ -3,9 +3,13 @@ from __future__ import annotations
 import argparse
 import json
 import os
+import sys
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Iterable
+
+_PROJECT_ROOT = Path(__file__).resolve().parents[2]
+sys.path.insert(0, str(_PROJECT_ROOT))
 
 import mlflow
 import numpy as np
@@ -32,13 +36,9 @@ class FeatureSelectionResult:
     dropped_due_to_correlation: list[str]
 
 
-DEFAULT_DB_CONFIG = {
-    "host": os.getenv("POSTGRES_HOST", "localhost"),
-    "port": int(os.getenv("POSTGRES_PORT", 5432)),
-    "dbname": os.getenv("POSTGRES_DB", "kkbox"),
-    "user": os.getenv("POSTGRES_USER", "bt4301"),
-    "password": os.getenv("POSTGRES_PASSWORD", "bt4301pass"),
-}
+from source.common.db import get_db_config
+
+DEFAULT_DB_CONFIG = get_db_config()
 
 
 def load_feature_store(db_config: dict = DEFAULT_DB_CONFIG) -> pd.DataFrame:
